refactor: 🎨 update src files
- New container directory for src files
- Rename src files to conform to the snake_case naming convention
- Update Customer class
