--- Customer.py
@@ -1,3 +1,0 @@
-class Customer:
-    loves: list[str] = []
-    dislike: list[str] = []
