--- Node.py
@@ -1,11 +1,0 @@
-class Node:
-    edges = []
-    name = ""
-    def listEdges(self):
-        for i in self.edges:
-            print(self.edges[i])
-
-class Edge:
-    edge = ()
-    def __init__(self, a, b):
-        self.edge(a, b)
